bot/pyproject.toml
[project]
name = "placebo-bot"
version = "0.1.0"
description = "Telegram bot for daily health check-ins"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
